# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Laravel",
        "contact": {
            "name": "hello@example.com",
            "email": "hello@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/customer-medication-purchase/store": {
            "post": {
                "tags": [
                    "CustomerTransactions"
                ],
                "summary": "Store new Customer Medication Purchas",
                "description": "Store new Customer Medication Purchas",
                "operationId": "purchaseMedication",
                "requestBody": {
                    "description": "Register new Customer Medication Purchas",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerMedicationPurchaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/customer-medication-return-request/store": {
            "post": {
                "tags": [
                    "CustomerTransactions"
                ],
                "summary": "Store new Customer Medication Return",
                "description": "Store new Customer Medication Return",
                "operationId": "returnMedication",
                "requestBody": {
                    "description": "Create new Customer Medication Return",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerMedicationReturnRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomerMedicationPurchaseRequest": {
                "title": "Customer Medication Purchase Request",
                "description": "Customer Medication Purchase Request body data",
                "required": [
                    "customer_id",
                    "medication_id",
                    "quantity"
                ],
                "properties": {
                    "quantity": {
                        "title": "quantity",
                        "description": "Quantity",
                        "example": 1
                    },
                    "remarks": {
                        "title": "remarks",
                        "description": "additional details about the transaction",
                        "type": "string",
                        "example": "remark"
                    },
                    "medication_id": {
                        "title": "medication_id",
                        "description": "Medication id",
                        "example": 1
                    },
                    "customer_id": {
                        "title": "customer_id",
                        "description": "Customer id",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "CustomerMedicationReturnRequest": {
                "title": "Customer Medication Return Request",
                "description": "Customer Medication Return Request body data",
                "required": [
                    "customer_id",
                    "medication_id",
                    "quantity"
                ],
                "properties": {
                    "quantity": {
                        "title": "quantity",
                        "description": "Quantity",
                        "example": 1
                    },
                    "remarks": {
                        "title": "remarks",
                        "description": "additional details about the transaction",
                        "type": "string",
                        "example": "remark"
                    },
                    "medication_id": {
                        "title": "medication_id",
                        "description": "Medication id",
                        "example": 1
                    },
                    "customer_id": {
                        "title": "customer_id",
                        "description": "Customer id",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "JsonResponseResource": {
                "title": "JsonResponseResource",
                "description": "Json Response Resource",
                "properties": {
                    "status": {
                        "title": "Status",
                        "description": "Status",
                        "type": "string"
                    },
                    "message": {
                        "title": "Message",
                        "description": "Message",
                        "type": "string"
                    },
                    "data": {
                        "title": "Data",
                        "description": "Data",
                        "nullable": true
                    },
                    "errors": {
                        "title": "Errors",
                        "description": "Errors",
                        "nullable": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "JsonResponseResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}